Split the citation-evidence markdown test into its own article CLI sidecar
