+++ src/cli/article/tests/mod.rs
@@ -1,5 +1,6 @@
 //! Article CLI test ownership modules.
 
+mod citation_evidence;
 mod exact_lookup;
 mod filters;
 mod help;
